requirements.txt
cryptography==3.4.8
lazyme==0.0.27
pycryptodome==3.18.0
PySide2==5.15.2.1
python_bcrypt==0.3.2
